gralloc: always just return 0 from check_for_compression() with no checks
In check_for_compression(), the value of the property ddk.set.afbc
was checked. We don't have it set, and neither do any stock firmwares
for any Exynos7870 devices, and the 7870 doesn't support AFBC (Arm
Frame Buffer Compression) so it can't be set to 1. If that property
is set to 0, check_for_compression() always returned 0. So, cut the
checks and simply always do that.

The reading of the value of ddk.set.afbc causes SELinux denials under 12,
so by doing this, the inconvenience of addressing them (properly with a
new proper property context) just for it is avoided. Also, the size of
the 64-bit gralloc.exynos5 blob is decreased by a fair amount.

Change-Id: I80809e57493c91e2a5233b4d555f73c6f82c6d3d